requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
joblib==1.4.2
numpy==1.24.4
pandas==2.0.3
pathlib==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
threadpoolctl==3.5.0
tzdata==2024.2
urllib3==2.2.3
